+++ src/req/PIDInfo.ts
@@ -1,0 +1,11 @@
+export interface PIDInfo
+{
+    pid : number;
+    ppid : number;
+    command : string;
+    arguments : Array<string>;
+    isPHAT : boolean;
+    isPHATRenderer : boolean;
+    isPHATMain : boolean;
+    url : string;
+}
