+++ src/req/getPIDInfo.ts
@@ -1,18 +1,11 @@
 import * as electron from "electron";
-const BrowserWindow = electron.remote.BrowserView;
+const BrowserWindow = electron.remote.BrowserWindow;
 
 
 const psNode = require("ps-node");
 
-export interface PIDInfo
-{
-    pid : number;
-    ppid : number;
-    command : string;
-    arguments : Array<string>;
-    isPHAT : boolean;
-    url : string;
-}
+import {PIDInfo} from "./PIDInfo";
+
 
 export async function getPIDInfo(pid : number) : Promise<Array<PIDInfo>>
 {
@@ -23,12 +16,14 @@ export async function getPIDInfo(pid : number) : Promise<Array<PIDInfo>>
                 reject(err);
             for(let i = 0; i != resultList.length; ++i)
             {
-                res.push(<PIDInfo>{
+                let info = <PIDInfo>{
                     pid : resultList[i].pid,
                     ppid : resultList[i].ppid,
                     command : resultList[i].command,
                     arguments : resultList[i].arguments
-                });
+                };
+
+                res.push(info);
             }
             resolve(res);
         });
